refactor(Repositories, DTOs, Services): 把实体的ID变成long类型，并采用雪花ID

+++ src/main/java/org/ivan/artshow/common/auth/AuthInterceptor.java
@@ -85,7 +85,7 @@ public class AuthInterceptor implements HandlerInterceptor {
 
         // 4. Validate and parse token
         try {
-            Integer userId = JwtUtils.parseToken(token);
+            Long userId = JwtUtils.parseToken(token);
             log.debug("Token validated successfully, user ID: {}", userId);
 
             // 5. Store in context
